making acf more robust

+++ gprot/acf/acf.py
@@ -7,7 +7,7 @@ def acf(x, y, maxlag=100):
     """Assumes regular sampling, zero-filled
     """
     cadence = np.median(np.diff(x))
-    maxlag_cad = maxlag/cadence
+    maxlag_cad = int(maxlag/cadence)
     ac = acor.function(y, maxlag_cad)
     lags = np.arange(len(ac))*cadence
     return lags, ac
